+++ packages/app/package.json
@@ -26,7 +26,6 @@
     "@visx/text": "1.3.0",
     "@visx/tooltip": "^1.3.0",
     "@visx/voronoi": "^1.0.0",
-    "@visx/xychart": "^1.4.0",
     "d3-scale": "^3.2.3",
     "date-fns": "^2.17.0",
     "geojson": "^0.5.0",
@@ -39,7 +38,6 @@
     "polished": "^4.1.0",
     "react": "^17.0.1",
     "react-dom": "^17.0.1",
-    "react-spring": "^8.0.27",
     "rehype-stringify": "^8.0.0",
     "remark-external-links": "^8.0.0",
     "remark-parse": "^9.0.0",
