--- packages/app/src/components-styled/sewer-chart-2/index.ts
@@ -1,1 +1,0 @@
-export * from './sewer-chart-2';
